Remove copied code, add extended classes where needed
