+++ core/org.osate.xtext.aadl2.ui/META-INF/MANIFEST.MF
@@ -30,11 +30,10 @@ Require-Bundle: org.eclipse.ui.editors;bundle-version="3.7.0",
  org.eclipse.xtext.xbase.lib,
  org.eclipse.xtend.lib,
  org.osate.pluginsupport;bundle-version="1.0.0",
- org.eclipse.jdt.core;bundle-version="3.13.0"
+ org.eclipse.jdt.core;bundle-version="3.13.0",
+ org.yakindu.base.xtext.utils.jface;bundle-version="3.5.9"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: de.itemis.utils.jface.viewers,
- de.itemis.xtext.utils.jface.viewers,
- org.osate.xtext.aadl2.ui,
+Export-Package: org.osate.xtext.aadl2.ui,
  org.osate.xtext.aadl2.ui.containers,
  org.osate.xtext.aadl2.ui.contentassist,
  org.osate.xtext.aadl2.ui.editor,
