--- core/org.osate.xtext.aadl2.ui/src/de/itemis/xtext/utils/jface/fieldassist/ICompletionProposalListener.java
@@ -1,28 +1,0 @@
-package de.itemis.xtext.utils.jface.fieldassist;
-
-/**
- * This interface is used to listen to additional notifications from a
- * {@link CompletionProposalAdapter}.
- * 
- * @author patrick.koenemann@itemis.de
- *
- */
-public interface ICompletionProposalListener {
-	/**
-	 * A completion proposal popup has been opened.
-	 * 
-	 * @param adapter
-	 *            the CompletionProposalAdapter which is adapting content proposal
-	 *            behavior to a control
-	 */
-	public void proposalPopupOpened(CompletionProposalAdapter adapter);
-
-	/**
-	 * A completion proposal popup has been closed.
-	 * 
-	 * @param adapter
-	 *            the CompletionProposalAdapter which is adapting content proposal
-	 *            behavior to a control
-	 */
-	public void proposalPopupClosed(CompletionProposalAdapter adapter);
-}
